python/experimental/requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
addict==2.4.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
bond==4.1.2
builtin-interfaces==1.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
composition-interfaces==1.2.1
ConfigArgParse==1.7
contourpy==1.3.1
control-msgs==4.8.0
controller-manager==2.50.0
controller-manager-msgs==2.50.0
cv-bridge==3.2.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
demo-nodes-py==0.20.5
deprecation==2.1.0
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.3
distro==1.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.57.0
gazebo-msgs==3.9.0
generate-parameter-library-py==0.4.0
geometry-msgs==4.8.0
graph-msgs==0.2.0
hightime==0.2.2
iae-webcam==0.0.0
idna==3.10
image-geometry==3.2.1
importlib_metadata==8.6.1
interactive-markers==2.3.2
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
laser-geometry==2.4.0
launch==1.0.8
launch-param-builder==0.1.1
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
message-filters==4.3.7
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
msgpack==1.1.0
nav-msgs==4.8.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nbformat==5.10.4
nest-asyncio==1.6.0
nidaqmx==1.0.2
numpy==2.2.1
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
open3d==0.19.0
opencv-python==4.10.0.84
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
PyDAQmx==1.4.7
Pygments==2.19.1
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
pyserial==3.5
python-can==4.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-qt-binding==1.1.2
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
realsense2-camera-msgs==4.55.1
referencing==0.36.2
requests==2.32.3
resource-retriever==3.1.3
retrying==1.3.4
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpds-py==0.24.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.4
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rtf-lds-driver==0.0.0
scikit-learn==1.6.1
scipy==1.15.2
scripts==3.9.0
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six==1.17.0
slam-toolbox==2.6.9
smclib==4.1.2
srdfdom==2.0.7
sros2==0.10.6
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
teleop-twist-keyboard==2.4.0
tenacity==9.0.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
theora-image-transport==2.5.3
threadpoolctl==3.5.0
topic-monitor==0.20.5
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==4.8.0
turtle-tf2-py==0.3.7
turtlesim==1.4.2
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.4.0
visualization-msgs==4.8.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xacro==2.0.13
zipp==3.21.0
